+++ src/session.rs
@@ -47,9 +47,10 @@ pub struct AuthenticatedRequestBuilder<'a, A: AuthMethod + 'a> {
 /// Owns a token and an underlying client.
 #[derive(Debug)]
 pub struct Session<Auth: AuthMethod> {
-    auth_method: Auth,
+    auth: Auth,
     client: Client,
-    cached_token: utils::ValueCache<Auth::TokenType>
+    cached_token: utils::ValueCache<Auth::TokenType>,
+    default_region: Option<String>
 }
 
 impl<'a, Auth: AuthMethod> AuthenticatedRequestBuilder<'a, Auth> {
@@ -105,9 +106,20 @@ impl<'a, Auth: AuthMethod + 'a> Session<Auth> {
     /// Create a new session with a given authentication plugin.
     pub fn new(auth_method: Auth) -> Session<Auth> {
         Session {
-            auth_method: auth_method,
+            auth: auth_method,
             client: utils::http_client(),
-            cached_token: utils::ValueCache::new(None)
+            cached_token: utils::ValueCache::new(None),
+            default_region: None
+        }
+    }
+
+    /// Create a new session with a given authentication plugin and region.
+    pub fn new_with_region(auth_method: Auth, region: String) -> Session<Auth> {
+        Session {
+            auth: auth_method,
+            client: utils::http_client(),
+            cached_token: utils::ValueCache::new(None),
+            default_region: Some(region)
         }
     }
 
@@ -117,12 +129,29 @@ impl<'a, Auth: AuthMethod + 'a> Session<Auth> {
         Ok(self.cached_token.get().unwrap())
     }
 
+    /// Get a reference to the authentication method in use.
+    pub fn auth_method(&self) -> &Auth {
+        &self.auth
+    }
+
+    /// Get a default (usually public) endpoint URL.
+    #[inline]
+    pub fn get_default_endpoint<S1: Into<String>>(&self, service_type: S1)
+            -> ApiResult<Url> {
+        self.auth.get_endpoint(service_type.into(),
+                                      None,
+                                      self.default_region.clone(),
+                                      &self)
+    }
+
     /// Get an endpoint URL.
-    pub fn get_endpoint(&self, service_type: &str,
-                        endpoint_interface: Option<&str>,
-                        region: Option<&str>) -> ApiResult<Url> {
-        self.auth_method.get_endpoint(service_type, endpoint_interface,
-                                      region, &self)
+    pub fn get_endpoint<S1, S2>(&self, service_type: S1,
+                                endpoint_interface: S2) -> ApiResult<Url>
+            where S1: Into<String>, S2: Into<String> {
+        self.auth.get_endpoint(service_type.into(),
+                                      Some(endpoint_interface.into()),
+                                      self.default_region.clone(),
+                                      &self)
     }
 
     /// A wrapper for HTTP request.
@@ -134,21 +163,9 @@ impl<'a, Auth: AuthMethod + 'a> Session<Auth> {
         }
     }
 
-    // Private and test-only
-
-    #[cfg(test)]
-    pub fn new_with_params(auth_method: Auth, client: Client,
-                           token: Auth::TokenType) -> Session<Auth> {
-        Session {
-            auth_method: auth_method,
-            client: client,
-            cached_token: utils::ValueCache::new(Some(token))
-        }
-    }
-
     fn refresh_token(&self) -> ApiResult<()> {
         self.cached_token.ensure_value(|| {
-            self.auth_method.get_token(&self.client)
+            self.auth.get_token(&self.client)
         })
     }
 }
@@ -173,29 +190,28 @@ pub struct ServiceApi<'a, Auth: AuthMethod + 'a, S> {
     session: &'a Session<Auth>,
     service_type: PhantomData<S>,
     endpoint_interface: Option<String>,
-    region: Option<String>,
     cached_endpoint: utils::ValueCache<Url>
 }
 
 impl<'a, Auth: AuthMethod + 'a, S: ServiceType> ServiceApi<'a, Auth, S> {
     /// Create a new API instance using the given session.
     pub fn new(session: &'a Session<Auth>) -> ServiceApi<'a, Auth, S> {
-        ServiceApi::new_with_endpoint_params(session, None, None)
-    }
-
-    /// Create a new API instance using the given session.
-    ///
-    /// This variant allows passing an endpoint type (defaults to public),
-    /// and region (defaults to any).
-    pub fn new_with_endpoint_params(session: &'a Session<Auth>,
-                                    endpoint_interface: Option<&str>,
-                                    region: Option<&str>)
-            -> ServiceApi<'a, Auth, S> {
         ServiceApi {
             session: session,
             service_type: PhantomData,
-            endpoint_interface: endpoint_interface.map(String::from),
-            region: region.map(String::from),
+            endpoint_interface: None,
+            cached_endpoint: utils::ValueCache::new(None)
+        }
+    }
+
+    /// Create a new API instance using the given session.
+    pub fn new_with_endpoint<S1>(session: &'a Session<Auth>,
+                                 endpoint_interface: S1)
+            -> ServiceApi<'a, Auth, S> where S1: Into<String> {
+        ServiceApi {
+            session: session,
+            service_type: PhantomData,
+            endpoint_interface: Some(endpoint_interface.into()),
             cached_endpoint: utils::ValueCache::new(None)
         }
     }
@@ -205,10 +221,11 @@ impl<'a, Auth: AuthMethod + 'a, S: ServiceType> ServiceApi<'a, Auth, S> {
     /// The resulting endpoint is cached on the current ServiceApi object.
     pub fn get_root_endpoint(&self, include_version: bool) -> ApiResult<Url> {
         try!(self.cached_endpoint.ensure_value(|| {
-            self.session.get_endpoint(
-                S::catalog_type(),
-                self.endpoint_interface.as_ref().map(String::as_str),
-                self.region.as_ref().map(String::as_str))
+            match self.endpoint_interface {
+                Some(ref s) => self.session.get_endpoint(S::catalog_type(),
+                                                         s.clone()),
+                None => self.session.get_default_endpoint(S::catalog_type())
+            }
         }));
 
         let endpoint = self.cached_endpoint.get().unwrap();
@@ -269,13 +286,42 @@ pub mod test {
     use hyper::status::StatusCode;
 
     use super::{ApiError, Session};
-    use super::super::auth::{Token, NoAuth, SimpleToken};
+    use super::super::auth::{Identity, Method, Token, NoAuth, SimpleToken};
+    use super::super::auth::identity::IdentityAuthMethod;
     use super::super::utils;
+
+    pub fn new_with_params<Auth: Method>(auth: Auth, cli: hyper::Client,
+                                         token: Auth::TokenType,
+                                         region: Option<&str>)
+            -> Session<Auth> {
+        Session {
+            auth: auth,
+            client: cli,
+            cached_token: utils::ValueCache::new(Some(token)),
+            default_region: region.map(From::from)
+        }
+    }
 
     pub fn new_session(token: &str) -> Session<NoAuth> {
         let token = SimpleToken(String::from(token));
-        Session::new_with_params(NoAuth::new("http://127.0.0.1/").unwrap(),
-                                 utils::http_client(), token)
+        new_with_params(NoAuth::new("http://127.0.0.1/").unwrap(),
+                        utils::http_client(), token, None)
+    }
+
+    fn session_with_identity(region: Option<&str>)
+            -> Session<IdentityAuthMethod> {
+        let id = Identity::new("http://127.0.2.1").unwrap()
+            .with_user("user", "pa$$w0rd", "example.com")
+            .with_project_scope("cool project", "example.com")
+            .create().unwrap();
+        let cli = hyper::Client::with_connector(MockCatalog::default());
+        let token = SimpleToken(String::from("abcdef"));
+        Session {
+            auth: id,
+            client: cli,
+            cached_token: utils::ValueCache::new(Some(token)),
+            default_region: region.map(From::from)
+        }
     }
 
     mock_connector!(MockHttp {
@@ -289,6 +335,50 @@ pub mod test {
                                {}"
     });
 
+    // Copied from keystone API reference.
+    const EXAMPLE_CATALOG_RESPONSE: &'static str = r#"
+    {
+        "catalog": [
+            {
+                "endpoints": [
+                    {
+                        "id": "39dc322ce86c4111b4f06c2eeae0841b",
+                        "interface": "public",
+                        "region": "RegionOne",
+                        "url": "http://localhost:5000"
+                    },
+                    {
+                        "id": "ec642f27474842e78bf059f6c48f4e99",
+                        "interface": "internal",
+                        "region": "RegionOne",
+                        "url": "http://localhost:5000"
+                    },
+                    {
+                        "id": "c609fc430175452290b62a4242e8a7e8",
+                        "interface": "admin",
+                        "region": "RegionOne",
+                        "url": "http://localhost:35357"
+                    }
+                ],
+                "id": "4363ae44bdf34a3981fde3b823cb9aa2",
+                "type": "identity",
+                "name": "keystone"
+            }
+        ],
+        "links": {
+            "self": "https://example.com/identity/v3/catalog",
+            "previous": null,
+            "next": null
+        }
+    }"#;
+
+    mock_connector!(MockCatalog {
+        "http://127.0.2.1" => String::from("HTTP/1.1 200 OK\r\n\
+                                            Server: Mock.Mock\r\n\
+                                            X-Subject-Token: abcdef\r\n
+                                            \r\n") + EXAMPLE_CATALOG_RESPONSE
+    });
+
     #[test]
     fn test_session_new() {
         let s = new_session("foo");
@@ -298,19 +388,13 @@ pub mod test {
     }
 
     #[test]
-    fn test_session_get_endpoint() {
-        let s = new_session("foo");
-        let e = s.get_endpoint("foo", None, None).unwrap();
-        assert_eq!(&e.to_string(), "http://127.0.0.1/");
-    }
-
-    #[test]
     fn test_session_request() {
         let cli = hyper::Client::with_connector(MockHttp::default());
         let s = Session {
-            auth_method: NoAuth::new("http://127.0.0.1/").unwrap(),
+            auth: NoAuth::new("http://127.0.0.1/").unwrap(),
             client: cli,
-            cached_token: utils::ValueCache::new(None)
+            cached_token: utils::ValueCache::new(None),
+            default_region: None
         };
 
         let mut resp = s.request(hyper::Post, "http://127.0.0.1/")
@@ -325,9 +409,10 @@ pub mod test {
     fn test_session_request_error() {
         let cli = hyper::Client::with_connector(MockHttp::default());
         let s = Session {
-            auth_method: NoAuth::new("http://127.0.0.2/").unwrap(),
+            auth: NoAuth::new("http://127.0.0.2/").unwrap(),
             client: cli,
-            cached_token: utils::ValueCache::new(None)
+            cached_token: utils::ValueCache::new(None),
+            default_region: None
         };
 
         let err = s.request(hyper::Post, "http://127.0.0.2/")
@@ -343,9 +428,10 @@ pub mod test {
     fn test_session_request_unchecked_error() {
         let cli = hyper::Client::with_connector(MockHttp::default());
         let s = Session {
-            auth_method: NoAuth::new("http://127.0.0.2/").unwrap(),
+            auth: NoAuth::new("http://127.0.0.2/").unwrap(),
             client: cli,
-            cached_token: utils::ValueCache::new(None)
+            cached_token: utils::ValueCache::new(None),
+            default_region: None
         };
 
         let mut resp = s.request(hyper::Post, "http://127.0.0.2/")
@@ -357,5 +443,52 @@ pub mod test {
         let mut s = String::new();
         resp.read_to_string(&mut s).unwrap();
         assert_eq!(&s, "{}");
+    }
+
+    #[test]
+    fn test_session_get_endpoint_no_region() {
+        let session = session_with_identity(None);
+
+        let e1 = session.get_default_endpoint("identity").unwrap();
+        assert_eq!(&e1.to_string(), "http://localhost:5000/");
+        let e2 = session.get_endpoint("identity", "admin").unwrap();
+        assert_eq!(&e2.to_string(), "http://localhost:35357/");
+
+        match session.get_default_endpoint("foo").err().unwrap() {
+            ApiError::EndpointNotFound(ref endp) =>
+                assert_eq!(endp, "foo"),
+            other => panic!("Unexpected {}", other)
+        };
+
+        match session.get_endpoint("identity", "unknown").err().unwrap() {
+            ApiError::EndpointNotFound(ref endp) =>
+                assert_eq!(endp, "identity"),
+            other => panic!("Unexpected {}", other)
+        };
+    }
+
+    #[test]
+    fn test_session_get_endpoint_with_region() {
+        let session = session_with_identity(Some("RegionOne"));
+
+        let e1 = session.get_endpoint("identity", "admin").unwrap();
+        assert_eq!(&e1.to_string(), "http://localhost:35357/");
+    }
+
+    #[test]
+    fn test_session_get_endpoint_with_region_fail() {
+        let session = session_with_identity(Some("unknown"));
+
+        match session.get_default_endpoint("identity").err().unwrap() {
+            ApiError::EndpointNotFound(ref endp) =>
+                assert_eq!(endp, "identity"),
+            other => panic!("Unexpected {}", other)
+        };
+
+        match session.get_endpoint("identity", "public").err().unwrap() {
+            ApiError::EndpointNotFound(ref endp) =>
+                assert_eq!(endp, "identity"),
+            other => panic!("Unexpected {}", other)
+        };
     }
 }